tests: Move integration tests to their own dir
To keep things tidy and ensure that the kernel tests just load the BPF
program.

Test Plan
=========

```
$ make test/integration GO=`which go`
```
